requirements.txt
Django==1.8.19
Pillow==5.1.0
django-filter==0.11.0
djangorestframework==3.1.3
psycopg2==2.7.4
vobject==0.9.5
django-adminactions==1.5
django-admin-bootstrapped==2.5.7
django-cors-headers==2.2.0
django-autoslug==1.9.3
gunicorn==19.8.1
